+++ KernelBench/level1/11_4D_tensor_matrix_multiplication.py
@@ -37,8 +37,8 @@ l = 256
 k = 768
 
 def get_inputs():
-    A = torch.rand(b, i, j, l, dtype=torch.bfloat16)
-    B = torch.rand(l, k, dtype=torch.bfloat16)
+    A = torch.rand(b, i, j, l, dtype=torch.half)
+    B = torch.rand(l, k, dtype=torch.half)
     return [A, B]
 
 def get_init_inputs():
