+++ KernelBench/level1/37_FrobeniusNorm_.py
@@ -30,7 +30,7 @@ dim1 = 512
 dim2 = 512
 
 def get_inputs():
-    x = torch.rand(batch_size, features, dim1, dim2, dtype=torch.bfloat16)
+    x = torch.rand(batch_size, features, dim1, dim2, dtype=torch.half)
     return [x]
 
 def get_init_inputs():
