+++ KernelBench/level1/55_conv_standard_2D__asymmetric_input__square_kernel.py
@@ -17,7 +17,7 @@ class Model(nn.Module):
     """
     def __init__(self, in_channels: int, out_channels: int, kernel_size: int, stride: int = 1, padding: int = 0, dilation: int = 1, groups: int = 1, bias: bool = False):
         super(Model, self).__init__()
-        self.conv2d = nn.Conv2d(in_channels, out_channels, (kernel_size, kernel_size), stride=stride, padding=padding, dilation=dilation, groups=groups, bias=bias, dtype=torch.bfloat16)
+        self.conv2d = nn.Conv2d(in_channels, out_channels, (kernel_size, kernel_size), stride=stride, padding=padding, dilation=dilation, groups=groups, bias=bias, dtype=torch.half)
         
     def forward(self, x: torch.Tensor) -> torch.Tensor:
         """
@@ -42,7 +42,7 @@ kernel_size = 3
 # asymmetric input: make width considerably larger than height
 
 def get_inputs():
-    x = torch.rand(batch_size, in_channels, height, width, dtype=torch.bfloat16)
+    x = torch.rand(batch_size, in_channels, height, width, dtype=torch.half)
     return [x]
 
 def get_init_inputs():
