+++ KernelBench/level1/78_conv_transposed_2D_asymmetric_input_asymmetric_kernel___padded__.py
@@ -15,7 +15,7 @@ class Model(nn.Module):
     """
     def __init__(self, in_channels: int, out_channels: int, kernel_size: tuple, stride: tuple = (1, 1), padding: tuple = (0, 0), bias: bool = False):
         super(Model, self).__init__()
-        self.conv_transpose2d = nn.ConvTranspose2d(in_channels, out_channels, kernel_size, stride=stride, padding=padding, bias=bias, dtype=torch.bfloat16)
+        self.conv_transpose2d = nn.ConvTranspose2d(in_channels, out_channels, kernel_size, stride=stride, padding=padding, bias=bias, dtype=torch.half)
         
     def forward(self, x: torch.Tensor) -> torch.Tensor:
         """
@@ -40,7 +40,7 @@ stride = (1, 1)
 padding = (1, 3)
 
 def get_inputs():
-    x = torch.rand(batch_size, in_channels, height, width, dtype=torch.bfloat16)
+    x = torch.rand(batch_size, in_channels, height, width, dtype=torch.half)
     return [x]
 
 def get_init_inputs():
