+++ KernelBench/level1/91_cumsum_reverse.py
@@ -21,7 +21,7 @@ input_shape = (32768,)
 dim = 1
 
 def get_inputs():
-    return [torch.rand(batch_size, *input_shape, dtype=torch.bfloat16)]
+    return [torch.rand(batch_size, *input_shape, dtype=torch.half)]
 
 def get_init_inputs():
     return [dim]
