+++ KernelBench/level2/13_ConvTranspose3d_Mean_Add_Softmax_Tanh_Scaling.py
@@ -13,8 +13,8 @@ class Model(nn.Module):
     """
     def __init__(self, in_channels, out_channels, kernel_size, stride, padding, scaling_factor):
         super(Model, self).__init__()
-        self.conv_transpose = nn.ConvTranspose3d(in_channels, out_channels, kernel_size, stride=stride, padding=padding, dtype=torch.bfloat16)
-        self.bias = nn.Parameter(torch.randn(1, out_channels, 1, 1, 1, dtype=torch.bfloat16))  # Broadcastable bias over channels
+        self.conv_transpose = nn.ConvTranspose3d(in_channels, out_channels, kernel_size, stride=stride, padding=padding, dtype=torch.half)
+        self.bias = nn.Parameter(torch.randn(1, out_channels, 1, 1, 1, dtype=torch.half))  # Broadcastable bias over channels
         self.scaling_factor = scaling_factor
 
     def forward(self, x):
@@ -37,7 +37,7 @@ padding = 1
 scaling_factor = 2.0
 
 def get_inputs():
-    return [torch.rand(batch_size, in_channels, depth, height, width, dtype=torch.bfloat16)]
+    return [torch.rand(batch_size, in_channels, depth, height, width, dtype=torch.half)]
 
 def get_init_inputs():
     return [in_channels, out_channels, kernel_size, stride, padding, scaling_factor]
